+++ index.html
@@ -24,6 +24,6 @@
 
     <div id="posts-container"></div>
     <button id="delete-completed-button">Удалить все выполненные задачи</button>
-    <script src="script.js" type="module"></script>
+    <script src="main.js" type="module"></script>
   </body>
 </html>
